refactor(app): 拆分主入口辅助界面
